publish

+++ Jenkinsfile
@@ -33,18 +33,13 @@ node
     		packageUrl,
     		iconUrl)
     		
+    	publishRelease()
+    		
     	// These should all be performed at the point where you've
 // checked out your sources on the slave. A 'git' executable
 // must be available.
 // Most typical, if you're not cloning into a sub directory
 
-bat "\"${tool 'Git'}\" rev-parse HEAD > GIT_COMMIT"
-//sh('git rev-parse HEAD > GIT_COMMIT')
-git_commit=readFile('GIT_COMMIT')
-// short SHA, possibly better for chat notifications, etc.
-short_commit=git_commit.take(6)
-echo short_commit
-    		
 
 /*
 
@@ -104,36 +99,41 @@ def updateModule(def id, def version, def platformVersion, def title, def descri
 
 def publishRelease()
 {
-	def userInput = input(id: 'userInput', message: 'Create release?') 
-	node {
-	echo "Compressing artifacts into one file"
-	dir('deploy')
+	// check for publish commit & master branch
+	bat "\"${tool 'Git'}\" git log -1 --pretty=%B > LAST_COMMIT_MESSAGE"
+	git_last_commit=readFile('LAST_COMMIT_MESSAGE')
+	
+	if (env.BRANCH_NAME == 'master' && git_last_commit == 'publish')
 	{
-		deleteDir()
-	}
+		//def userInput = input(id: 'userInput', message: 'Create release?') 
+		echo "Compressing artifacts into one file"
+		dir('deploy')
+		{
+			deleteDir()
+		}
+		
+		zip dir: '', glob: '', zipFile: 'deploy\\artifacts.zip'
+		
 	
-	zip dir: '', glob: '', zipFile: 'deploy\\artifacts.zip'
+		//bat "${env.Utils}\\github-release info -u VirtoCommerce -r vc-module-jenkinssample"
+		bat "${env.Utils}\\github-release release --user VirtoCommerce --repo vc-module-jenkinssample --tag v1.0 --name v1.0"
+		bat "${env.Utils}\\github-release upload --user VirtoCommerce --repo vc-module-jenkinssample --tag v1.0 --name v1.0 --file \"deploy\\artifacts.zip\""
 	
-
-	//bat "${env.Utils}\\github-release info -u VirtoCommerce -r vc-module-jenkinssample"
-	bat "${env.Utils}\\github-release release --user VirtoCommerce --repo vc-module-jenkinssample --tag v1.0 --name v1.0"
-	bat "${env.Utils}\\github-release upload --user VirtoCommerce --repo vc-module-jenkinssample --tag v1.0 --name v1.0 --file \"deploy\\artifacts.zip\""
-
-	/*
-	zip -r artifacts.zip artifacts_folder
-	
-	echo "Exporting token and enterprise api to enable github-release tool"
-	export GITHUB_TOKEN=$$$$$$$$$$$$
-	export GITHUB_API=https://git.{your domain}.com/api/v3 # needed only for enterprise
-	
-	echo "Deleting release from github before creating new one"
-	/path_to_bin/./github-release delete --user ${GITHUB_ORGANIZATION} --repo ${GITHUB_REPO} --tag ${VERSION_NAME}
-	
-	echo "Creating a new release in github"
-	/path_to_bin/./github-release release --user ${GITHUB_ORGANIZATION} --repo ${GITHUB_REPO} --tag ${VERSION_NAME} --name "${VERSION_NAME}"
-	
-	echo "Uploading the artifacts into github"
-	/path_to_bin/./github-release upload --user ${GITHUB_ORGANIZATION} --repo ${GITHUB_REPO} --tag ${VERSION_NAME} --name "${PROJECT_NAME}-${VERSION_NAME}.zip" --file artifacts.zip
-	*/
+		/*
+		zip -r artifacts.zip artifacts_folder
+		
+		echo "Exporting token and enterprise api to enable github-release tool"
+		export GITHUB_TOKEN=$$$$$$$$$$$$
+		export GITHUB_API=https://git.{your domain}.com/api/v3 # needed only for enterprise
+		
+		echo "Deleting release from github before creating new one"
+		/path_to_bin/./github-release delete --user ${GITHUB_ORGANIZATION} --repo ${GITHUB_REPO} --tag ${VERSION_NAME}
+		
+		echo "Creating a new release in github"
+		/path_to_bin/./github-release release --user ${GITHUB_ORGANIZATION} --repo ${GITHUB_REPO} --tag ${VERSION_NAME} --name "${VERSION_NAME}"
+		
+		echo "Uploading the artifacts into github"
+		/path_to_bin/./github-release upload --user ${GITHUB_ORGANIZATION} --repo ${GITHUB_REPO} --tag ${VERSION_NAME} --name "${PROJECT_NAME}-${VERSION_NAME}.zip" --file artifacts.zip
+		*/
 	}
 }
